+++ software/rocketlogger/sensor/tsl4531.h
@@ -70,20 +70,20 @@ enum TSL4531_range {
 /*
  * API FUNCTIONS
  */
-int TSL4531_init(uint8_t);
-void TSL4531_close(uint8_t);
-int TSL4531_read(uint8_t);
-int32_t TSL4531_getValue(uint8_t, uint8_t);
+int TSL4531_init(int);
+void TSL4531_close(int);
+int TSL4531_read(int);
+int32_t TSL4531_getValue(int, int);
 
-int TSL4531_setRange(uint8_t, int);
-int TSL4531_getRange(uint8_t);
+int TSL4531_setRange(int, int);
+int TSL4531_getRange(int);
 
 /*
  * Helper FUNCTIONS
  */
-uint8_t TSL4531_getID(void);
-int TSL4531_setParameters(uint8_t);
-int TSL4531_sendRange(uint8_t, int);
-int TSL4531_getIndex(uint8_t);
+int TSL4531_getID(void);
+int TSL4531_setParameters(int);
+int TSL4531_sendRange(int, int);
+int TSL4531_getIndex(int);
 
 #endif /* SENSOR_TSL4531_H_ */
